+++ psdi_data_conversion/testing/utils.py
@@ -21,7 +21,8 @@ from unittest.mock import patch
 import py
 import pytest
 
-from psdi_data_conversion.constants import CONVERTER_OB, GLOBAL_LOG_FILENAME, LOG_NONE, OUTPUT_LOG_EXT
+from psdi_data_conversion.constants import (CONVERTER_AUTOCHAIN, CONVERTER_OB, GLOBAL_LOG_FILENAME, LOG_NONE,
+                                            OUTPUT_LOG_EXT)
 from psdi_data_conversion.converter import run_converter, run_converter_chain
 from psdi_data_conversion.converters.openbabel.converter import COORD_GEN_KEY, COORD_GEN_QUAL_KEY
 from psdi_data_conversion.database import get_converter_info, get_format_info
@@ -161,7 +162,7 @@ class ConversionTestSpec:
     compatible_with_library: bool = True
     """Whether or not this test spec is compatible with being run through the Python library, default True"""
 
-    compatible_with_cla: bool = True
+    compatible_with_cli: bool = True
     """Whether or not this test spec is compatible with being run through the command-line interface, default True"""
 
     compatible_with_gui: bool = True
@@ -370,7 +371,7 @@ def _run_single_test_conversion_with_library(test_spec: SingleConversionTestSpec
 
         # If we're provided a target format and converter, turn this into a path
         if ("path" not in conversion_kwargs and test_spec.to_format is not None and
-                test_spec.converter_name is not None):
+                test_spec.converter_name is not None and test_spec.converter_name != CONVERTER_AUTOCHAIN):
             conversion_kwargs["path"] = [(get_converter_info(test_spec.converter_name),
                                           get_format_info(test_spec.to_format))]
         elif test_spec.to_format is not None:
@@ -422,7 +423,7 @@ def _run_single_test_conversion_with_library(test_spec: SingleConversionTestSpec
             pytest.fail(callback_msg)
 
 
-def run_test_conversion_with_cla(test_spec: ConversionTestSpec):
+def run_test_conversion_with_cli(test_spec: ConversionTestSpec):
     """Runs a test conversion or series thereof through the command-line interface.
 
     Parameters
@@ -438,13 +439,13 @@ def run_test_conversion_with_cla(test_spec: ConversionTestSpec):
                 print(f"Skipping single test spec {single_test_spec}")
                 continue
             print(f"Running single test spec: {single_test_spec}")
-            _run_single_test_conversion_with_cla(test_spec=single_test_spec,
+            _run_single_test_conversion_with_cli(test_spec=single_test_spec,
                                                  input_dir=input_dir,
                                                  output_dir=output_dir)
             print(f"Success for test spec: {single_test_spec}")
 
 
-def _run_single_test_conversion_with_cla(test_spec: SingleConversionTestSpec,
+def _run_single_test_conversion_with_cli(test_spec: SingleConversionTestSpec,
                                          input_dir: str,
                                          output_dir: str):
     """Runs a single test conversion through the command-line interface.
@@ -472,18 +473,21 @@ def _run_single_test_conversion_with_cla(test_spec: SingleConversionTestSpec,
         stdouterr = py.io.StdCaptureFD(in_=False)
 
         if test_spec.expect_success:
-            run_converter_through_cla(filename=qualified_in_filename,
-                                      to_format=test_spec.to_format,
-                                      from_format=test_spec.from_format,
-                                      name=test_spec.converter_name,
-                                      input_dir=input_dir,
-                                      output_dir=output_dir,
-                                      log_file=os.path.join(output_dir, test_spec.log_filename),
-                                      **test_spec.conversion_kwargs)
-            success = True
+            try:
+                run_converter_through_cli(filename=qualified_in_filename,
+                                          to_format=test_spec.to_format,
+                                          from_format=test_spec.from_format,
+                                          name=test_spec.converter_name,
+                                          input_dir=input_dir,
+                                          output_dir=output_dir,
+                                          log_file=os.path.join(output_dir, test_spec.log_filename),
+                                          **test_spec.conversion_kwargs)
+                success = True
+            except SystemExit:
+                success = False
         else:
             with pytest.raises(SystemExit) as exc_info:
-                run_converter_through_cla(filename=qualified_in_filename,
+                run_converter_through_cli(filename=qualified_in_filename,
                                           to_format=test_spec.to_format,
                                           from_format=test_spec.from_format,
                                           name=test_spec.converter_name,
@@ -505,6 +509,13 @@ def _run_single_test_conversion_with_cla(test_spec: SingleConversionTestSpec,
         # Reset stdout and stderr capture
         stdouterr.done()
 
+    # If failed, print any stdout and stderr
+    if not success:
+        if stdout:
+            print(stdout)
+        if stderr:
+            print(stderr, file=sys.stderr)
+
     # Compile output info for the test and call the callback function if one is provided
     if test_spec.callback:
         test_info = ConversionTestInfo(run_type="cli",
@@ -520,7 +531,7 @@ def _run_single_test_conversion_with_cla(test_spec: SingleConversionTestSpec,
             pytest.fail(callback_msg)
 
 
-def run_converter_through_cla(filename: str,
+def run_converter_through_cli(filename: str,
                               to_format: str,
                               name: str,
                               input_dir: str,